+++ x/crosschain/keeper/rate_limiter_flags_test.go
@@ -1,7 +1,6 @@
 package keeper_test
 
 import (
-	"strings"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -10,7 +9,6 @@ import (
 	"github.com/zeta-chain/zetacore/pkg/coin"
 	keepertest "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/testutil/sample"
-	"github.com/zeta-chain/zetacore/x/crosschain/keeper"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	fungibletypes "github.com/zeta-chain/zetacore/x/fungible/types"
 )
@@ -24,7 +22,7 @@ func createForeignCoinAndAssetRate(
 	decimals uint32,
 	coinType coin.CoinType,
 	rate sdk.Dec,
-) (fungibletypes.ForeignCoins, *types.AssetRate) {
+) (fungibletypes.ForeignCoins, types.AssetRate) {
 	// create foreign coin
 	foreignCoin := sample.ForeignCoins(t, zrc20Addr)
 	foreignCoin.Asset = asset
@@ -33,13 +31,13 @@ func createForeignCoinAndAssetRate(
 	foreignCoin.CoinType = coinType
 
 	// create corresponding asset rate
-	assetRate := &types.AssetRate{
-		ChainId:  foreignCoin.ForeignChainId,
-		Asset:    strings.ToLower(foreignCoin.Asset),
-		Decimals: foreignCoin.Decimals,
-		CoinType: foreignCoin.CoinType,
-		Rate:     rate,
-	}
+	assetRate := sample.CustomAssetRate(
+		foreignCoin.ForeignChainId,
+		foreignCoin.Asset,
+		foreignCoin.Decimals,
+		foreignCoin.CoinType,
+		rate,
+	)
 
 	return foreignCoin, assetRate
 }
@@ -59,7 +57,7 @@ func TestKeeper_GetRateLimiterFlags(t *testing.T) {
 	require.Equal(t, flags, r)
 }
 
-func TestKeeper_GetRateLimiterRateList(t *testing.T) {
+func TestKeeper_GetRateLimiterAssetRateList(t *testing.T) {
 	k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 
 	// create test flags
@@ -67,7 +65,7 @@ func TestKeeper_GetRateLimiterRateList(t *testing.T) {
 	zrc20GasAddr := sample.EthAddress().Hex()
 	zrc20ERC20Addr1 := sample.EthAddress().Hex()
 	zrc20ERC20Addr2 := sample.EthAddress().Hex()
-	flags := types.RateLimiterFlags{
+	testflags := types.RateLimiterFlags{
 		Rate: sdk.NewUint(100),
 		Conversions: []types.Conversion{
 			{
@@ -85,8 +83,14 @@ func TestKeeper_GetRateLimiterRateList(t *testing.T) {
 		},
 	}
 
+	// asset rates not found before setting flags
+	flags, assetRates, found := k.GetRateLimiterAssetRateList(ctx)
+	require.False(t, found)
+	require.Equal(t, types.RateLimiterFlags{}, flags)
+	require.Nil(t, assetRates)
+
 	// set flags
-	k.SetRateLimiterFlags(ctx, flags)
+	k.SetRateLimiterFlags(ctx, testflags)
 
 	// add gas coin
 	gasCoin, gasAssetRate := createForeignCoinAndAssetRate(t, zrc20GasAddr, "", chainID, 18, coin.CoinType_Gas, sdk.NewDec(1))
@@ -101,48 +105,8 @@ func TestKeeper_GetRateLimiterRateList(t *testing.T) {
 	zk.FungibleKeeper.SetForeignCoins(ctx, erc20Coin2)
 
 	// get rates
-	assetRates := k.GetRateLimiterAssetRateList(ctx)
-	require.EqualValues(t, []*types.AssetRate{gasAssetRate, erc20AssetRate1, erc20AssetRate2}, assetRates)
-}
-
-func TestBuildAssetRateMapFromList(t *testing.T) {
-	// define asset rate list
-	assetRates := []*types.AssetRate{
-		{
-			ChainId:  1,
-			Asset:    "eth",
-			Decimals: 18,
-			CoinType: coin.CoinType_Gas,
-			Rate:     sdk.NewDec(1),
-		},
-		{
-			ChainId:  1,
-			Asset:    "usdt",
-			Decimals: 6,
-			CoinType: coin.CoinType_ERC20,
-			Rate:     sdk.NewDec(2),
-		},
-		{
-			ChainId:  2,
-			Asset:    "btc",
-			Decimals: 8,
-			CoinType: coin.CoinType_Gas,
-			Rate:     sdk.NewDec(3),
-		},
-	}
-
-	// build asset rate map
-	gasAssetRateMap, erc20AssetRateMap := keeper.BuildAssetRateMapFromList(assetRates)
-
-	// check length
-	require.Equal(t, 2, len(gasAssetRateMap))
-	require.Equal(t, 1, len(erc20AssetRateMap))
-	require.Equal(t, 1, len(erc20AssetRateMap[1]))
-
-	// check gas asset rate map
-	require.EqualValues(t, assetRates[0], gasAssetRateMap[1])
-	require.EqualValues(t, assetRates[2], gasAssetRateMap[2])
-
-	// check erc20 asset rate map
-	require.EqualValues(t, assetRates[1], erc20AssetRateMap[1]["usdt"])
+	flags, assetRates, found = k.GetRateLimiterAssetRateList(ctx)
+	require.True(t, found)
+	require.Equal(t, testflags, flags)
+	require.EqualValues(t, []types.AssetRate{gasAssetRate, erc20AssetRate1, erc20AssetRate2}, assetRates)
 }